compose_2/requirements.txt
requests
mysql-connector-python
psycopg2-binary
pymongo
redis
pytz
